+++ backend/endpoints/__init__.py
@@ -1,5 +1,6 @@
 """Import all files with endpoints here. `app.py` imports this file to easily import all endpoints."""
 
-from . import mat_endpoints
-from . import frontend_endpoints
-from . import import_export
+from . import creating
+from . import reading
+from . import updating
+# from . import deleting
